Test 74 passed: test_method_names_become_symbols
Lesson learned and answer to "THINK ABOUT IT":
Why do we use "Symbol.all_symbols.map { |x| x.to_s } ? Why do this to test whether there is a "test_method_names_become_symbols" symbol?

We do this because even if there was no "test_method_names_become_symbols" symbol, if we wrote:
assert_equal true, Symbol.all_symbols.include?(:test_method_names_become_symbols)
That very act would create the symbol we were looking for. So we use Symbol.all_symbols to get an array of all the symbols currenly in Ruby's symbol table, use the .map block to convert each to a string value which goes in our array, then test for the string version of the symbol we're looking for. This way, if the symbol doesn't exist before testing for it, we won't accidentally create it.